refactor(donation-receipt): migrate DonationReceiptContext to Zustand store
Replace the DonationReceiptContext provider with a Zustand store, matching
the established store + useInitializeX + StoreInitializer pattern.

- Add donationReceiptStore with state, actions, imperative getters, and
  selectReceiptData / selectOperation selectors
- Hydrate from sessionStorage via useInitializeDonationReceipt, registered
  in StoreInitializer; persist on each mutation under the same key
- Update consumers to use selectors for reactive reads and store actions
  for mutations, reducing re-renders vs the context
- Remove DonationReceiptProvider from _app and delete the context file
- Drop unused getVerificationDate

No behavioral change to receipt verification, issuance, or donor updates.
