+++ pages/_app.tsx
@@ -27,7 +27,6 @@ import materialTheme from '../src/theme/themeStyles';
 import { PlanetCashProvider } from '../src/features/common/Layout/PlanetCashContext';
 import { PayoutsProvider } from '../src/features/common/Layout/PayoutsContext';
 import { NextIntlClientProvider } from 'next-intl';
-import { DonationReceiptProvider } from '../src/features/common/Layout/DonationReceiptContext';
 import { StoreInitializer } from '../src/features/common/StoreInitializer/StoreInitializer';
 
 const Layout = dynamic(() => import('../src/features/common/Layout'), {
@@ -190,11 +189,7 @@ const PlanetWeb = ({
                 <PlanetCashProvider>
                   <PayoutsProvider>
                     <Layout>
-                      <BulkCodeProvider>
-                        <DonationReceiptProvider>
-                          {pageContent}
-                        </DonationReceiptProvider>
-                      </BulkCodeProvider>
+                      <BulkCodeProvider>{pageContent}</BulkCodeProvider>
                     </Layout>
                   </PayoutsProvider>
                 </PlanetCashProvider>
