+++ src/features/user/DonationReceipt/DonorContactManagement.tsx
@@ -13,13 +13,17 @@ import AddAddress from '../Settings/EditProfile/AddressManagement/AddAddress';
 import EditAddress from '../Settings/EditProfile/AddressManagement/EditAddress';
 import { ADDRESS_ACTIONS } from '../../../utils/addressManagement';
 import { useApi } from '../../../hooks/useApi';
-import { useDonationReceiptContext } from '../../common/Layout/DonationReceiptContext';
 import DonorContactForm from './microComponents/DonorContactForm';
 import { transformProfileToDonorView } from './transformers'; // TODO: remove for production
 import { validateOwnership } from './DonationReceiptValidator';
 import EditPermissionDenied from './microComponents/EditPermissionDenied';
 import { RECEIPT_STATUS } from './donationReceiptTypes';
-import { useUserStore, useErrorHandlingStore } from '../../../stores';
+import {
+  useUserStore,
+  useErrorHandlingStore,
+  useDonationReceiptStore,
+  selectOperation,
+} from '../../../stores';
 
 type IndividualProfile = {
   firstname: string;
@@ -35,8 +39,6 @@ type CompanyProfile = {
 type ProfileData = IndividualProfile | CompanyProfile;
 
 const DonorContactManagement = () => {
-  const { updateDonorAndAddress, email, tinIsRequired, getOperation } =
-    useDonationReceiptContext();
   const t = useTranslations('DonationReceipt');
   const router = useRouter();
   const { putApiAuthenticated } = useApi();
@@ -52,12 +54,18 @@ const DonorContactManagement = () => {
   );
   // store: state
   const userProfile = useUserStore((state) => state.userProfile);
+  const email = useDonationReceiptStore((state) => state.email);
+  const tinIsRequired = useDonationReceiptStore((state) => state.tinIsRequired);
+  const operation = useDonationReceiptStore(selectOperation);
   // store: action
   const setUserProfile = useUserStore((state) => state.setUserProfile);
   const setErrors = useErrorHandlingStore((state) => state.setErrors);
+  const updateDonorAndAddress = useDonationReceiptStore(
+    (state) => state.updateDonorAndAddress
+  );
 
   const isOwner = validateOwnership(email, userProfile?.email);
-  if (!isOwner && getOperation() !== RECEIPT_STATUS.ISSUE)
+  if (!isOwner && operation !== RECEIPT_STATUS.ISSUE)
     return <EditPermissionDenied />;
 
   // Navigate back to the verification page
